+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The following changes have been completed, and will be included in the next tagg
 
 ### New
 
+* Archetypes / Utility Types (new module!)
+  - added `AttributeNames` utility type
 * Basic Types
   - added `validateArrayOf()`
   - added `validateNonEmptyArray()`